chewie: A package for downloading GEDI data.

The gedi package provides a set of functions for downloading and carrying out simple pre processing of GEDI data. See below for a list of these functions.

Managing Credentials and Cache

chewie_creds

Set up and manage NASA Earthdata credentials

chewie_health_check

Quick diagnostics and health checks for the package cache and credentials.

chewie_setup_cache

Set up and manage the package cache.

chewie_unset_cache

Remove the CHEWIE_PARQUET_CACHE environment variable from the .Renviron file.

chewie_get_cache

Get the path to the GEDI cache directory.

chewie_clear_find_cache

Clear the find gedi cache.

chewie_clear_h5_temp_cache

Clear the h5 temp cache.

chewie_clear_parquet_cache

Clear the parquet gedi cache.

Downloading GEDI Data

find_gedi

Find GEDI data in a given area for a given time period.

grab_gedi

Download GEDI data and store in the cache.

Reading GEDI Data

collect_gedi

Collect GEDI data into a data frame from parquet files in the cache.

Visualising GEDI Data

chewie_show

Visualise GEDI data on a map.

See also

Author

Maintainer: Hugh Graham hugh.graham@permianglobal.com (ORCID)